requirements.txt
transformers>=4.40.0
torch>=2.0.0
torchvision
pillow
requests
